general improvements

+++ Software/STM/src/stm32/BSP/Thermometer/Thermometer.cpp
@@ -4,7 +4,7 @@
 
 #include <BSP/Thermometer.hpp>
 
-float BSP2::Thermometer::get_temperature() const
+double BSP2::Thermometer::get_temperature() const
 {
     return temperature_;
 }
