+++ Software/STM/src/stm32/BSP/Thermometer/include/BSP/Thermometer.hpp
@@ -1,22 +1,23 @@
 /**
-  * Thermometer.hpp
-  *
-  *  Created on: 26-07-2025
-  *  @author: Paweł Warzecha
-  */
+ * Thermometer.hpp
+ *
+ *  Created on: 26-07-2025
+ *  @author: Paweł Warzecha
+ */
 
 #pragma once
 
-namespace BSP2{
+namespace BSP2
+{
 
 class Thermometer
 {
 public:
-    [[nodiscard]] float get_temperature() const;
+    [[nodiscard]] double get_temperature() const;
     void update();
 
 private:
-    float temperature_{};
+    double temperature_{};
 };
 
-}
+}  // namespace BSP2
